reenable hexlify unit-test and fix broken hexlify

+++ src/common_methods/object_methods.cpp
@@ -40,7 +40,7 @@ hexlify(std::string &outputString,
         if(bytestream[i - 1] < 16) {
             stream << "0";
         }
-        stream << std::hex << bytestream[i - 1];
+        stream << std::hex << static_cast<int>(bytestream[i - 1]);
     }
 
     outputString = stream.str();
